+++ LobotomyCorporationMods.DontChatMe/UiComponents/ConnectionState.cs
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+
+namespace LobotomyCorporationMods.DontChatMe.UiComponents
+{
+    public enum ConnectionState
+    {
+        Disabled,
+        Disconnected,
+        Connecting,
+        Connected,
+    }
+}